requirements.txt
numpy~=1.25.0
pandas==2.0.3
setuptools~=65.5.0
openpyxl==3.1.2
matplotlib~=3.7.1
scipy~=1.11.1
PyYAML~=6.0